SY-2749: Move channel rename service to channel/service
