+++ console/src/channel/services/external.ts
@@ -7,7 +7,10 @@
 // License, use of this software will be governed by the Apache License, Version 2.0,
 // included in the file licenses/APL.txt.
 
+export * from "@/channel/services/channelNameUtils";
+export * from "@/channel/services/channelRenameService";
 export * from "@/channel/services/link";
 export * from "@/channel/services/ontology";
 export * from "@/channel/services/palette";
+export * from "@/channel/services/channelRenameModal";
 export * from "@/channel/services/Toolbar";
